+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-21T10:38:46.427Z")
+    created: Optional[datetime] = parse_datetime("2026-02-21T14:44:02.644Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")