+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHZWFXDWZ8FSK85C9XC69W44"
+        "01KJ0AH0REWK5YDWC3CNK07VNJ"
     )
     r"""The unique identifier of the data source"""