+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHZWFXDZH8VZ1FZ5H7YMZDMM"
+        "01KJ0AH0RGWSTFDYTSA92V42DW"
     )
     r"""The unique identifier of the data source"""