+++ packages/orq-rc/RELEASES.md
@@ -5919,3 +5919,13 @@ Based on:
 - [python v4.4.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.24 - packages/orq-rc
+
+## 2026-02-21 14:46:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.25 - packages/orq-rc